analytics-service/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
pandas==2.2.3
numpy==2.2.3
scikit-learn==1.6.1
matplotlib==3.10.1
pytest==8.3.5
httpx==0.28.1
python-dotenv==1.0.1
